improve output regarding this command
